all fns refactored
