+++ console/src/components/Services/Actions/Permissions/utils.js
@@ -1,8 +1,8 @@
+import { findActionPermission } from '../utils';
 import {
   getCreateActionPermissionQuery,
   getDropActionPermissionQuery,
-} from '../../../Common/utils/v1QueryUtils';
-import { findActionPermission } from '../utils';
+} from '../../../../metadata/queryUtils';
 
 export const getActionPermissionQueries = (
   permissionEdit,
