+++ console/src/components/Services/Data/TableBrowseRows/ViewActions.js
@@ -13,7 +13,6 @@ import {
   getSelectQuery,
   getFetchManualTriggersQuery,
   getDeleteQuery,
-  getRunSqlQuery,
 } from '../../../Common/utils/v1QueryUtils';
 import { COUNT_LIMIT } from '../constants';
 import {
@@ -21,6 +20,7 @@ import {
   dataSource,
   findTableFromRel,
 } from '../../../../dataSources';
+import { getRunSqlQuery } from '../../../../metadata/queryUtils';
 
 /* ****************** View actions *************/
 const V_SET_DEFAULTS = 'ViewTable/V_SET_DEFAULTS';
