+++ console/src/components/Services/Settings/About/About.tsx
@@ -9,9 +9,9 @@ import globals from '../../../../Globals';
 import styles from '../Settings.scss';
 import requestAction from '../../../../utils/requestAction';
 import { showErrorNotification } from '../../Common/Notification';
-import { getRunSqlQuery } from '../../../Common/utils/v1QueryUtils';
 import { versionGT } from '../../../../helpers/versionUtils';
 import { ReduxState, ConnectInjectedProps } from '../../../../types';
+import { getRunSqlQuery } from '../../../../metadata/queryUtils';
 
 type AboutState = {
   consoleAssetVersion?: string;
